requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
crypto==1.4.1
Flask==1.1.2
gcloud==0.18.3
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jws==0.1.3
MarkupSafe==1.1.1
multidict==4.7.6
Naked==0.1.31
oauth2client==4.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.8
pyee==7.0.2
Pyrebase4==4.3.0
python-dotenv==0.14.0
python-jwt==2.0.1
PyYAML==5.3.1
requests==2.24.0
requests-toolbelt==0.9.1
rsa==4.6
schedule==0.6.0
shellescape==3.8.1
six==1.15.0
slack==0.0.2
slackclient==1.2.1
slackeventsapi==2.0.0
urllib3==1.25.9
websocket-client==0.57.0
Werkzeug==1.0.1
yarl==1.4.2
